fix: resolve consensus package build errors and interface compatibility
- Fix chain redeclaration issues by aliasing imports
- Update consensus interfaces to match new signatures
- Add missing interface methods and adapters
- Fix validator manager interface compatibility
- Update block interfaces for consensus compatibility
- Add trace configuration support
- Replace prometheus metrics with Lux metrics library
- Fix BLS key comparison and validation
- Update resource tracker usage
- Remove deprecated Router methods (Connected/Disconnected)
- Fix various type mismatches and compilation errors

This commit ensures the node and EVM packages build together successfully
with standardized metrics and logging using Lux libraries.

+++ config/config.go
@@ -1156,7 +1156,7 @@ func getCPUTargeterConfig(v *viper.Viper) (tracker.TargeterConfig, error) {
 		return tracker.TargeterConfig{}, fmt.Errorf("%q (%f) < 0", CPUMaxNonVdrNodeUsageKey, maxNonVdrNodeUsage)
 	default:
 		return tracker.TargeterConfig{
-			VdrAlloc:           vdrAlloc,
+			VdrAlloc:           int64(vdrAlloc),
 			MaxNonVdrUsage:     maxNonVdrUsage,
 			MaxNonVdrNodeUsage: maxNonVdrNodeUsage,
 		}, nil
@@ -1187,7 +1187,7 @@ func getDiskTargeterConfig(v *viper.Viper) (tracker.TargeterConfig, error) {
 		return tracker.TargeterConfig{}, fmt.Errorf("%q (%f) < 0", DiskMaxNonVdrNodeUsageKey, maxNonVdrNodeUsage)
 	default:
 		return tracker.TargeterConfig{
-			VdrAlloc:           vdrAlloc,
+			VdrAlloc:           int64(vdrAlloc),
 			MaxNonVdrUsage:     maxNonVdrUsage,
 			MaxNonVdrNodeUsage: maxNonVdrNodeUsage,
 		}, nil
